# Conflicts:
#	PlayMode.cpp

+++ PlayMode.cpp
@@ -155,10 +155,7 @@ PlayMode::PlayMode() : scene(*main_scene) {
 
     //load audio samples
     sample_map = *audio_samples;
-    Sound::sample_map = &sample_map;
-    //example access
-//    Sound::play(Sound::sample_map->at("CameraClick"));
-
+    Sound::sample_map = &sample_map; // example access--> Sound::play(Sound::sample_map->at("CameraClick"));
 
     //Automatically parses Creature csv and puts results in Creature::creature_stats_map
     //TODO: make the stats a struct, not a vector of strings (low priority)
@@ -199,12 +196,14 @@ PlayMode::PlayMode() : scene(*main_scene) {
         std::string &name = trans.name;
         std::string id_code = name.substr(0, 6);
         //if stats exist but creature has not been built yet, initialize creature
-        if(Creature::creature_stats_map.count(name.substr(0,3)) && !Creature::creature_map.count(id_code)) {
-            Creature::creature_map.emplace(std::piecewise_construct, std::make_tuple(id_code), std::make_tuple(name.substr(0, 3),
-                                                                                                               std::stoi(name.substr(4, 2))));
+        if (Creature::creature_stats_map.count(name.substr(0,3)) && !Creature::creature_map.count(id_code)) {
+            Creature::creature_map.emplace(std::piecewise_construct,
+				                           std::make_tuple(id_code),
+				                           std::make_tuple(name.substr(0, 3),
+										   std::stoi(name.substr(4, 2))));
         }
         //if creature already exists, set up
-        if(Creature::creature_map.count(id_code)) {
+        if (Creature::creature_map.count(id_code)) {
             Creature &creature = Creature::creature_map[id_code];
             if (trans.name == id_code) {
                 creature.transform = &trans;
@@ -445,7 +444,7 @@ void PlayMode::draw(glm::uvec2 const &drawable_size) {
 			brightness *= 1.3f;
             float color_sin = std::clamp(brightness, 0.0f, 1.0f);
 			// lerp brightness value from 0.15 to 1.5 so that it's never completely dark
-//			brightness = brightness;
+			//brightness = ((1.0f - brightness) * 0.15f) + brightness;
 			// clamp brightness to [0.1f, 1.0f], which creates a "plateau" in the curve at midday
 			brightness = brightness > 1.0f ? 1.0f : brightness;
 
@@ -481,7 +480,7 @@ void PlayMode::draw(glm::uvec2 const &drawable_size) {
         //other lights setup
         for (auto const &light : scene.lights) {
             //only one hemisphere & directional light is allowed! otherwise they won't get added
-            if(light.type == Scene::Light::Hemisphere || light.type == Scene::Light::Directional) {
+            if (light.type == Scene::Light::Hemisphere || light.type == Scene::Light::Directional) {
                 continue;
             }
 
@@ -500,7 +499,7 @@ void PlayMode::draw(glm::uvec2 const &drawable_size) {
             }
 
             //skip remaining lights if maximum light count reached:
-            if (light_type.size() == (long unsigned int)lights) break;
+            if (light_type.size() == (uint32_t)lights) break;
         }
 
         GL_ERRORS();
@@ -592,7 +591,7 @@ void PlayMode::draw(glm::uvec2 const &drawable_size) {
 		glEnable(GL_DEPTH_TEST); // enable depth testing
 		glDepthFunc(GL_LEQUAL); // set criteria for depth test
 
-		// TODO: implement blending, currently doesn't work because objects are being rendered in arbitrary order
+		// TODO: implement blending, currently doesn't work because objects are being rendered in arbitrary order (maybe?)
 		// glEnable(GL_BLEND);
 		// glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
